+++ HISTORY.rst
@@ -7,6 +7,8 @@
   - Enforce same origin and HTTP only redirection when breaking from an AJAX
     content container and during normal single page app navigation.
   - Git ignore `__pycache__` folders.
+  - Make the Smart link condition fields `inclusion` and `operator` fully
+    translatable.
 
 4.6.11 (2025-11-08)
 ===================